requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.13.0
timm>=0.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
opencv-python>=4.8.0
albumentations>=1.3.0
requests>=2.28.0
python-dotenv
vastai-sdk
kaggle>=1.5.16 